+++ addon/components/paper-checkbox.js
@@ -8,7 +8,7 @@ const { inject, assert } = Ember;
 export default BaseFocusable.extend(RippleMixin, ProxiableMixin, ColorMixin, {
   tagName: 'md-checkbox',
   classNames: ['md-checkbox', 'md-default-theme'],
-  classNameBindings: ['checked:md-checked'],
+  classNameBindings: ['value:md-checked'],
 
   /* Ripple Overrides */
   rippleContainerSelector: '.md-container',
@@ -21,7 +21,7 @@ export default BaseFocusable.extend(RippleMixin, ProxiableMixin, ColorMixin, {
 
   constants: inject.service(),
 
-  checked: false,
+  value: false,
 
   didInitAttrs() {
     this._super(...arguments);
@@ -30,7 +30,7 @@ export default BaseFocusable.extend(RippleMixin, ProxiableMixin, ColorMixin, {
 
   click() {
     if (!this.get('disabled')) {
-      this.sendAction('onChange', !this.get('checked'));
+      this.sendAction('onChange', !this.get('value'));
     }
   },
 
@@ -43,6 +43,6 @@ export default BaseFocusable.extend(RippleMixin, ProxiableMixin, ColorMixin, {
   },
 
   processProxy() {
-    this.sendAction('onChange', !this.get('checked'));
+    this.sendAction('onChange', !this.get('value'));
   }
 });